feat(user): move <UserTabs> to <UserProfile>
