+++ src/components/index.tsx
@@ -69,4 +69,3 @@ export * from './Throw404'
 export * from './Transaction'
 export * from './TranslationsProvider'
 export * from './UserDigest'
-export * from './UserProfile'
